+++ deepparse/data_validation/__init__.py
@@ -1,0 +1,2 @@
+# pylint: disable=wildcard-import
+from .data_validation import *