+++ examples/parse_addresses.py
@@ -26,6 +26,13 @@ address_parser = AddressParser(model_type="bpemb", device=0)
 # We can now parse some addresses
 parsed_addresses = address_parser(test_data[0:300])
 
+# When parsing addresses, some data quality tests are applied to the dataset.
+# First, it validates that no addresses to parse are empty.
+# Second, it validates that no addresses are whitespace-only.
+# The next two lines are rising a DataError.
+address_parser("")  # Raise an error
+address_parser(" ")  # Raise an error
+
 # We can also put our parsed address into a pandas dataframe for analysis
 # You can choose the fields to use or use the default one
 fields = ["StreetNumber", "StreetName", "Municipality", "Province", "PostalCode"]