+++ tests/parser/test_address_parser.py
@@ -9,6 +9,7 @@ from unittest.mock import patch, MagicMock
 import torch
 from torch import device
 
+from deepparse.data_error import DataError
 from deepparse.parser import FormattedParsedAddress, formatted_parsed_address
 from deepparse.parser.address_parser import AddressParser
 from tests.parser.base import AddressParserPredictTestCase
@@ -1526,6 +1527,40 @@ class AddressParserTest(AddressParserPredictTestCase):
             actual = address_parser.get_formatted_model_name()
             expected = "BPEmbAttention"
             self.assertEqual(expected, actual)
+
+    @patch("deepparse.parser.address_parser.BPEmbEmbeddingsModel")
+    @patch("deepparse.parser.address_parser.BPEmbVectorizer")
+    @patch("deepparse.parser.address_parser.bpemb_data_padding")
+    def test_givenEmptyData_whenParse_raiseDataError(
+        self, embeddings_model_mock, vectorizer_model_mock, data_padding_mock
+    ):
+        empty_data = ["an address", ""]
+        another_empty_address = ""
+        with patch("deepparse.parser.address_parser.BPEmbSeq2SeqModel") as model_mock:
+            self.mock_predictions_vectors(model_mock)
+            address_parser = AddressParser(model_type=self.a_bpemb_model_type, device=self.a_cpu_device)
+            with self.assertRaises(DataError):
+                address_parser(empty_data)
+
+            with self.assertRaises(DataError):
+                address_parser(another_empty_address)
+
+    @patch("deepparse.parser.address_parser.BPEmbEmbeddingsModel")
+    @patch("deepparse.parser.address_parser.BPEmbVectorizer")
+    @patch("deepparse.parser.address_parser.bpemb_data_padding")
+    def test_givenWhiteSpaceOnlyData_whenParse_raiseDataError(
+        self, embeddings_model_mock, vectorizer_model_mock, data_padding_mock
+    ):
+        whitespace_data = ["an address", " "]
+        another_whitespace_address = " "
+        with patch("deepparse.parser.address_parser.BPEmbSeq2SeqModel") as model_mock:
+            self.mock_predictions_vectors(model_mock)
+            address_parser = AddressParser(model_type=self.a_bpemb_model_type, device=self.a_cpu_device)
+            with self.assertRaises(DataError):
+                address_parser(whitespace_data)
+
+            with self.assertRaises(DataError):
+                address_parser(another_whitespace_address)
 
 
 if __name__ == "__main__":